+++ src/actions/minting/index.ts
@@ -1,4 +1,4 @@
-export * from "./getApprovedTokenAmount";
+export * from "./getTokenSpendAllowance";
 export * from "./isMinterApprovedToSpendTokens";
 export * from "./mintNetworkDomain";
 export * from "./isNetworkDomainAvailable";
